# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tender API",
    "version": "0.1.0",
    "description": "Any-chain crypto checkout. Buyers pay from any chain; merchants settle in one asset on Monad.\n\nMerchant routes need **Authorize → `tk_live_…`** (create one with `npm run merchant:create`). Public routes are what the buyer checkout calls. Money is always a decimal **string**; the wire is **snake_case**."
  },
  "components": {
    "securitySchemes": {
      "merchantKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Merchant API key (tk_live_…)"
      },
      "platformKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dashboard platform key (tp_…). On /v1 routes, also send `X-Tender-Merchant: mer_…`. Server-side only."
      }
    }
  },
...
  "paths": {
...
    "/public/invoices/{token}": {
      "get": {
        "tags": [
          "Public (checkout)"
        ],
        "summary": "An invoice, as the buyer may see it",
        "description": "No merchant email, settlement address, internal id or reference. `minimum` is in USD.",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "DETECTED",
                        "SETTLED",
                        "OVERPAID",
                        "UNDERPAID",
                        "EXPIRED",
                        "CANCELLED",
                        "NEEDS_RECOVERY"
                      ]
                    },
                    "amount_expected": {
                      "type": "string",
                      "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                    },
                    "currency": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "merchant_name": {
                      "type": "string"
                    },
                    "addresses": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "chain": {
                            "type": "string",
                            "minLength": 1
                          },
                          "address": {
                            "type": "string"
                          },
                          "minimum": {
                            "type": "string",
                            "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                          }
                        },
                        "required": [
                          "chain",
                          "address"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "redirect_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "token",
                    "status",
                    "amount_expected",
                    "currency",
                    "expires_at",
                    "merchant_name",
                    "addresses"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/public/invoices/{token}/events": {
      "get": {
        "tags": [
          "Public (checkout)"
        ],
        "summary": "Live status (server-sent events)",
        "description": "One `data:` frame per status change, each an InvoiceEvent. The current status is sent first; the stream closes on a terminal status. Swagger UI cannot display a stream — try it with `curl -N`.",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "An SSE stream of InvoiceEvent frames",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/public/invoices/{token}/submit-tx": {
      "post": {
        "tags": [
          "Public (checkout)"
        ],
        "summary": "Buyer submits their tx hash to speed up detection",
        "security": [],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "tx_hash": {
                    "type": "string",
                    "description": "The transfer's transaction hash"
                  }
                },
                "required": [
                  "tx_hash"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "accepted": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "accepted"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed — `fields` names each bad field",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Not found (or not yours)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/public/chains": {
      "get": {
        "tags": [
          "Public (checkout)"
        ],
        "summary": "Supported chains with measured minimums (USD) and settlement time",
        "description": "503 until the worker has measured minimums — a minimum is never guessed.",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "minLength": 1
                      },
                      "name": {
                        "type": "string"
                      },
                      "asset": {
                        "type": "string"
                      },
                      "minimum": {
                        "type": "string",
                        "pattern": "^(0|[1-9]\\d*)(\\.\\d{1,18})?$"
                      },
                      "estimated_settlement": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "asset",
                      "minimum",
                      "estimated_settlement"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Not ready yet (or, on the assistant, the model is unavailable)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/public/fx": {
      "get": {
        "tags": [
          "Public (checkout)"
        ],
        "summary": "Display-currency rates (units of each currency per 1 USD)",
        "description": "Reference rates for SHOWING amounts in another currency. They never price an invoice or move money, and update about once a day. 503 when none are available: a rate is never guessed.",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "base": {
                      "type": "string",
                      "const": "USD"
                    },
                    "as_of": {
                      "type": "string",
                      "description": "When the rates were published (ISO 8601). They update about once a day."
                    },
                    "rates": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      },
                      "description": "Units of each currency per 1 USD, as decimal strings"
                    }
                  },
                  "required": [
                    "base",
                    "as_of",
                    "rates"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "503": {
            "description": "Not ready yet (or, on the assistant, the model is unavailable)",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "fields": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  }
}